function that prints all the element of a list

+++ 0x12-singly_linked_lists/0-main.c
@@ -27,13 +27,13 @@ int main(void)
     new->next = head;
     head = new;
     n = print_list(head);
-    printf("-> %zu elements\n", n);
+    printf("-> %lu elements\n", (long unsigned int)n);
 
     printf("\n");
     free(new->str);
     new->str = NULL;
     n = print_list(head);
-    printf("-> %zu elements\n", n);
+    printf("-> %lu elements\n", (long unsigned int)n);
 
     free(new);
     return (0);
